+++ tests/integration/DisputeTest.php
@@ -19,7 +19,7 @@ class DisputeTest extends Setup
         ]);
     }
 
-    public function testCreate_evidenceDocument()
+    public function createSampleDocument()
     {
     }
 
@@ -28,11 +28,8 @@ class DisputeTest extends Setup
         $result = Braintree\Transaction::sale([
             'amount' => '100.00',
             'creditCard' => [
-                'number' => '4023898493988028',
-                'expirationDate' => '05/2009',
-            ],
-            'options' => [
-                'skipAdvancedFraudChecking' => true
+                'number' => Braintree\Test\CreditCardNumbers::$disputes['Chargeback'],
+                'expirationDate' => '12/2019',
             ]
         ]);
         return $result->transaction->disputes[0];
@@ -47,27 +44,59 @@ class DisputeTest extends Setup
 
         $updatedDispute = $this->gateway->dispute()->find($dispute->id);
 
-        $this->assertEquals(Braintree\Dispute::ACCEPTED, $dispute->status);
+        $this->assertEquals(Braintree\Dispute::ACCEPTED, $updatedDispute->status);
     }
 
     public function testAccept_errors_whenDisputeNotOpen()
     {
+        $result = $this->gateway->dispute()->accept("wells_dispute");
+        $error = $result->errors->forKey('dispute')->errors[0];
+
+        $this->assertFalse($result->success);
+        $this->assertEquals(Braintree\Error\Codes::DISPUTE_CAN_ONLY_ACCEPT_OPEN_DISPUTE, $error->code);
+        $this->assertEquals("Disputes can only be accepted when they are in an Open state", $error->message);
     }
 
     public function testAccept_raisesError_whenDisputeNotFound()
     {
+        $this->setExpectedException('Braintree\Exception\NotFound', 'dispute with id "invalid-id" not found');
+        $this->gateway->dispute()->accept("invalid-id");
     }
 
     public function testAddFileEvidence_addsEvidence()
     {
+        // $disputeId = $this->createSampleDispute()->id;
+        // $documentId = createSampleDocument()->id;
+        //
+        // $result = $this->gateway->dispute()->addFileEvidence($disputeId, $documentId);
+        //
+        // $this->assertTrue($result->success);
+        //
+        // $updatedDispute = $this->gateway->dispute()->find($disputeId);
+        //
+        // $this->assertEquals($result->evidence->id, $updatedDispute->evidence[0]->id);
     }
 
     public function testAddFileEvidence_raisesError_whenDisputeNotFound()
     {
+        // $this->setExpectedException('Braintree\Exception\NotFound', 'dispute with id "unknown_dispute_id" not found');
+        // $this->gateway->dispute()->addFileEvidence("unknown_dispute_id", "unknown_file_id");
     }
 
     public function testAddFileEvidence_raisesError_whenDisputeNotOpen()
     {
+        // $disputeId = $this->createSampleDispute()->id;
+        // $documentId = $this->createSampleDocument()->id;
+        //
+        // $this->gateway->dispute()->accept($disputeId);
+        //
+        // $result = $this->gateway->dispute()->addFileEvidence($disputeId, $documentId);
+        // $error = $result->errors->forKey('dispute')->errors[0];
+        //
+        // $this->assertFalse($result->success);
+        // $this->assertEquals(Braintree\Error\Codes::DISPUTE_CAN_ONLY_ADD_EVIDENCE_TO_OPEN_DISPUTE, $error->code);
+        // $this->assertEquals("Evidence can only be attached to disputes that are in an Open state", $error->message);
+        // $this->assertEquals(
     }
 
     public function testAddFileEvidence_returnsError_whenIncorrectDocumentKind()
@@ -76,30 +105,79 @@ class DisputeTest extends Setup
 
     public function testAddTextEvidence_addsTextEvidence()
     {
+        $disputeId = $this->createSampleDispute()->id;
+
+        $result = $this->gateway->dispute()->addTextEvidence($disputeId, "text evidence");
+        $evidence = $result->evidence;
+
+        $this->assertTrue($result->success);
+        $this->assertEquals("text evidence", $evidence->comment);
+        $this->assertNotNull($evidence->createdAt);
+        $this->assertRegExp('/^\w{16,}$/', $evidence->id);
+        $this->assertNull($evidence->sentToProcessorAt);
+        $this->assertNull($evidence->url);
     }
 
     public function testAddTextEvidence_raisesError_whenDisputeNotFound()
     {
+        $this->setExpectedException('Braintree\Exception\NotFound', 'dispute with id "unknown_dispute_id" not found');
+        $dispute = $this->gateway->dispute()->addTextEvidence("unknown_dispute_id", "text evidence");
     }
 
     public function testAddTextEvidence_raisesError_whenDisputeNotOpen()
     {
+        $disputeId = $this->createSampleDispute()->id;
+
+        $this->gateway->dispute()->accept($disputeId);
+
+        $result = $this->gateway->dispute()->addTextEvidence($disputeId, "text evidence");
+        $error = $result->errors->forKey('dispute')->errors[0];
+
+        $this->assertFalse($result->success);
+        $this->assertEquals(Braintree\Error\Codes::DISPUTE_CAN_ONLY_ADD_EVIDENCE_TO_OPEN_DISPUTE, $error->code);
+        $this->assertEquals("Evidence can only be attached to disputes that are in an Open state", $error->message);
     }
 
     public function testAddTextEvidence_showsNewRecord_inFind()
     {
+        $disputeId = $this->createSampleDispute()->id;
+
+        $evidence = $this->gateway->dispute()->addTextEvidence($disputeId, "text evidence")->evidence;
+
+        $refreshedDispute = $this->gateway->dispute()->find($disputeId);
+        $refreshedEvidence = $refreshedDispute->evidence[0];
+
+        $this->assertEquals($evidence->id, $refreshedEvidence->id);
+        $this->assertEquals($evidence->comment, $refreshedEvidence->comment);
     }
 
     public function testFinalize_changesDisputeStatus_toDisputed()
     {
+        $disputeId = $this->createSampleDispute()->id;
+
+        $result = $this->gateway->dispute()->finalize($disputeId);
+
+        $this->assertTrue($result->success);
+
+        $updatedDispute = $this->gateway->dispute()->find($disputeId);
+
+        $this->assertEquals(Braintree\Dispute::DISPUTED, $updatedDispute->status);
     }
 
     public function testFinalize_errors_whenDisputeNotOpen()
     {
+        $result = $this->gateway->dispute()->finalize("wells_dispute");
+        $error = $result->errors->forKey('dispute')->errors[0];
+
+        $this->assertFalse($result->success);
+        $this->assertEquals(Braintree\Error\Codes::DISPUTE_CAN_ONLY_FINALIZE_OPEN_DISPUTE, $error->code);
+        $this->assertEquals("Disputes can only be finalized when they are in an Open state", $error->message);
     }
 
     public function testFinalize_raisesError_whenDisputeNotFound()
     {
+        $this->setExpectedException('Braintree\Exception\NotFound', 'dispute with id "invalid-id" not found');
+        $result = $this->gateway->dispute()->finalize("invalid-id");
     }
 
     public function testFind_returnsDispute_withGivenId()
@@ -119,92 +197,34 @@ class DisputeTest extends Setup
         $this->gateway->dispute()->find("invalid-id");
     }
 
-    public function testtRemoveEvidence_removesEvidencFromTheDisupute()
+    public function testRemoveEvidence_removesEvidenceFromTheDisupute()
     {
+        $disputeId = $this->createSampleDispute()->id;
+        $evidenceId = $this->gateway->dispute()->addTextEvidence($disputeId, "text evidence")->evidence->id;
+
+        $result = $this->gateway->dispute()->removeEvidence($disputeId, $evidenceId);
+
+        $this->assertTrue($result->success);
     }
 
-    public function testtRemoveEvidence_raisesError_whenDisputeOrEvidenceNotFound()
+    public function testRemoveEvidence_raisesError_whenDisputeOrEvidenceNotFound()
     {
+        $this->setExpectedException('Braintree\Exception\NotFound', "evidence with id \"unknown_evidence_id\" for dispute with id \"unknown_dispute_id\" not found");
+        $this->gateway->dispute()->removeEvidence("unknown_dispute_id", "unknown_evidence_id");
     }
 
-    public function testtRemoveEvidence_errors_whenDisputeNotOpen()
+    public function testRemoveEvidence_errors_whenDisputeNotOpen()
     {
-    }
+        $disputeId = $this->createSampleDispute()->id;
+        $evidenceId = $this->gateway->dispute()->addTextEvidence($disputeId, "text evidence")->evidence->id;
 
-    public function testUpdate_withValidationErrors()
-    {
-        // $customer = Braintree\Customer::createNoValidate();
-        // $address = Braintree\Address::createNoValidate([
-        //     'customerId' => $customer->id,
-        //     'streetAddress' => '1 E Main St'
-        // ]);
-        // $result = Braintree\Address::update(
-        //     $customer->id,
-        //     $address->id,
-        //     [
-        //         'countryName' => 'Invalid States of America'
-        //     ]
-        // );
-        // $this->assertFalse($result->success);
-        // $countryErrors = $result->errors->forKey('address')->onAttribute('countryName');
-        // $this->assertEquals(Braintree\Error\Codes::ADDRESS_COUNTRY_NAME_IS_NOT_ACCEPTED, $countryErrors[0]->code);
-    }
+        $this->gateway->dispute()->accept($disputeId);
 
-    public function testUpdate_withValidationErrors_onCountry()
-    {
-        // $customer = Braintree\Customer::createNoValidate();
-        // $address = Braintree\Address::createNoValidate([
-        //     'customerId' => $customer->id,
-        //     'streetAddress' => '1 E Main St'
-        // ]);
-        // $result = Braintree\Address::update(
-        //     $customer->id,
-        //     $address->id,
-        //     [
-        //         'countryCodeAlpha2' => 'MU',
-        //         'countryCodeAlpha3' => 'MYT'
-        //     ]
-        // );
-        // $this->assertFalse($result->success);
-        // $countryErrors = $result->errors->forKey('address')->onAttribute('base');
-        // $this->assertEquals(Braintree\Error\Codes::ADDRESS_INCONSISTENT_COUNTRY, $countryErrors[0]->code);
-    }
+        $result = $this->gateway->dispute()->removeEvidence($disputeId, $evidenceId);
+        $error = $result->errors->forKey('dispute')->errors[0];
 
-
-    public function testUpdateNoValidate()
-    {
-        // $customer = Braintree\Customer::createNoValidate();
-        // $createdAddress = Braintree\Address::createNoValidate([
-        //     'customerId' => $customer->id,
-        //     'firstName' => 'Old First',
-        //     'lastName' => 'Old Last',
-        //     'company' => 'Old Company',
-        //     'streetAddress' => '1 E Old St',
-        //     'extendedAddress' => 'Apt Old',
-        //     'locality' => 'Old Chicago',
-        //     'region' => 'Old Region',
-        //     'postalCode' => 'Old Postal',
-        //     'countryName' => 'United States of America'
-        // ]);
-        // $address = Braintree\Address::updateNoValidate($customer->id, $createdAddress->id, [
-        //     'firstName' => 'New First',
-        //     'lastName' => 'New Last',
-        //     'company' => 'New Company',
-        //     'streetAddress' => '1 E New St',
-        //     'extendedAddress' => 'Apt New',
-        //     'locality' => 'New Chicago',
-        //     'region' => 'New Region',
-        //     'postalCode' => 'New Postal',
-        //     'countryName' => 'Mexico'
-        // ]);
-        // $this->assertEquals('New First', $address->firstName);
-        // $this->assertEquals('New Last', $address->lastName);
-        // $this->assertEquals('New Company', $address->company);
-        // $this->assertEquals('1 E New St', $address->streetAddress);
-        // $this->assertEquals('Apt New', $address->extendedAddress);
-        // $this->assertEquals('New Chicago', $address->locality);
-        // $this->assertEquals('New Region', $address->region);
-        // $this->assertEquals('New Postal', $address->postalCode);
-        // $this->assertEquals('Mexico', $address->countryName);
+        $this->assertFalse($result->success);
+        $this->assertEquals(Braintree\Error\Codes::DISPUTE_CAN_ONLY_REMOVE_EVIDENCE_FROM_OPEN_DISPUTE, $error->code);
+        $this->assertEquals("Evidence can only be removed from disputes that are in an Open state", $error->message);
     }
 }